Se integra API de Backend para la Creacion de Plan de Entrenamiento

+++ app/src/main/java/com/sportapp_grupo1/ui/PlanAlimentacionCreate.kt
@@ -83,7 +83,7 @@ class PlanAlimentacionCreate : Fragment() {
                     viernes = viernes,
                     sabado = sabado,
                     domingo = domingo,
-                    numero_semanas = semanas
+                    numero_semanas = semanas.toInt()
                 )
                 if (viewModel.addNewPlanAlimentacion(newPlan)) {
                     showMessage("El nuevo Plan Alimentacion se registró correctamente.")
